ULE-StarterKit-v1.0/fwtool/requirements.txt
pyserial
click
colorama
